Move tests from master to the right folder
