--- integration-tests/it-common/src/main/resources-binary/samples/test_tika_config_path/test.html
@@ -1,15 +1,0 @@
-<html
-    xmlns:o="urn:schemas-microsoft-com:office:office"
-    xmlns="http://www.w3.org/TR/REC-html40">
-
-<head>
-  <title>Test Tika title</title>
-</head>
-
-<body>
-<p>This is an example of HTML</p>
-<div />
-<p>This is a separate line</p>
-</body>
-
-</html>
