--- integration-tests/it-common/src/main/resources-binary/samples/test_tika_config_path/test.xhtml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE html>
-<html
-  xmlns:o="urn:schemas-microsoft-com:office:office"
-  xmlns="http://www.w3.org/TR/REC-html40">
-
-<head>
-  <title>Test Tika title</title>
-</head>
-
-<body>
-<p>This is an example of XHTML</p>
-<div />
-<p>This is a separate line</p>
-</body>
-
-</html>
