+++ integration-tests/src/test/resources-binary/samples/test_tika_config_path/test.html
@@ -1,0 +1,15 @@
+<html
+    xmlns:o="urn:schemas-microsoft-com:office:office"
+    xmlns="http://www.w3.org/TR/REC-html40">
+
+<head>
+  <title>Test Tika title</title>
+</head>
+
+<body>
+<p>This is an example of HTML</p>
+<div />
+<p>This is a separate line</p>
+</body>
+
+</html>
